resize hashtable if load goes over 0.75
To keep a balanced laod factor, the table will be resized when the load is greater or equal to 0.75. This corresponds to the java implementation. The provided implementation still needs some cleanup: it is the simplest way I could implement it. Contains duplicate code and handling of reinsert could use some improvement.

refers: #3